fixup: Refactor `Program` to not be static.
