+++ Cli-Shared/Delegates.cs
